+++ util.h
@@ -7,8 +7,8 @@
 // conversions
 int a2d(char ch);
 char a2i(char ch, char **src, int base, int *nump);
-void ui2a(unsigned int num, unsigned int base, char *bf);
-void i2a(int num, char *bf);
+void ui2a(uint64_t num, unsigned int base, char *bf);
+void i2a(int64_t num, char *bf);
 
 // memory
 void *memset(void *s, int c, size_t n);
